test: Add comprehensive tests for integrate_sin and integrate_cos
This commit isolates and improves tests for trigonometric integration functions:
- Extracted disorganized and duplicated integrate_sin/integrate_cos tests from test_Calculus.py
- Created a dedicated Tests/test_TrigIntegration.py module
- Added robust assertions validating bounds including negative, zero-width, fractional periods, and reversed paths
- Cleaned test_Calculus.py of redundant imports and broken artifacts from previous manual merges
